Fixed data types in getAttributes() of dataFrameDao and sqlDao

+++ tests/testthat/test-sqlDao.R
@@ -10,7 +10,7 @@ test_that("sqlDao", {
     expect_false(is.dao(list()))
 
     expect_equal(dao$getAttributes(),
-                 c(title='character', value='double'))
+                 c(title='character', value='numeric'))
     expect_equal(dao$getData(),
                  data.frame(id=1:5, title=letters[1:5], value=2 * 1:5, stringsAsFactors = FALSE))
     expect_equal(dao$getRecord(3),
